Ошибка #34377
Не понятно возникновение ошибки:
" Фактическое количество прайс листов меньше трех" в отчете №701

+++ src/ReportSystem/ByOffers/OffersReport.cs
@@ -64,7 +64,7 @@ namespace Inforoom.ReportSystem
 			_sourcePriceCode = _priceCode;
 			_customerFirmName = GetSupplierName(_priceCode);
 
-			CheckPriceActual(_sourcePriceCode);
+			CheckPriceActual((uint)_sourcePriceCode);
 			GetOffers(_SupplierNoise);
 			CheckSupplierCount();
